+++ R/deseq_fc_ecdf.R
@@ -1,4 +1,4 @@
-#' Plot the ecdf for DESeq2 log2 Fold Changes
+#' Plot the ECDF for DESeq2 log2(Fold Changes)
 #'
 #' @description This functions will take DESeq2 results as a `data.frame` and
 #' plot the ecdf for the input `gene.lists`.
@@ -10,9 +10,8 @@
 #' `gene.lists = list("Background" = c("gene1", "gene2"), "Target" = c("gene2",
 #' "gene3"), "Overlap" = c("gene2"))`
 #'
-#' This function will also perform statistical testing using the the
-#' `twosamples` package if `plot.hist` is TRUE. The output will be saved to a
-#' PDF if an `output.filename` is provided.
+#' This function will also perform statistical testing if `plot.hist` is TRUE.
+#' The output will be saved to a PDF if an `output.filename` is provided.
 #'
 #' Users can define the groups that are to be compared in the statistical test
 #' using the `null.name` and `target.name` arguments. The names must be found
@@ -22,7 +21,7 @@
 #' This functions returns:
 #'
 #' * `$plot`: The ECDF plot
-#' * `$stats`: The twosamples results object
+#' * `$stats`: The stats results object
 #'
 #' @param res The DESeq2 results dataframe
 #' @param gene.lists A nest list of gene names. Example:
@@ -36,18 +35,14 @@
 #' @param x.lims The xlimits range
 #' @param stats.test The statistic test to use. Options: KS, Kuiper, DTS, CVM,
 #' AD, Wass
-#' @param nboots The number of iterations to run the simulation
+#' @param alternative The alternative hypothesis to test. Options: greater, less, two.sided
 #' @param null.name The name in the gene.list to use as the null for ecdf plots
-#' @param target.name The name in the gene.list to use as the target for ecdf
-#' @param keep.boots Whether to keep the results from the boot strapping
-#' analysis
-#' @param plot.hist Whether to plot the histogram of dstats derived from boost
-#' strap analysis. Keep.boots must be TRUE.
+#' @param target.name The name in the gene.list to use as the target for ecdf plots
 #' @param height Plot height in inches
 #' @param width Plot width in inches
 #' @param dpi The dpi resolution for the figure
 #'
-#' @return A ggplot object of the ecdf
+#' @return A ggplot object for the ECDF plot
 #' @export
 #'
 #' @examplesIf interactive()
@@ -82,24 +77,21 @@
 #' stats.test = "KS",
 #' factor.order = c("Background", "mer7m8"),
 #' null.name = "Background",
-#' target.name = "mer7m8",
-#' nboots = 100)
+#' target.name = "mer7m8")
 deseq_fc_ecdf <- function(res,
                           gene.lists,
-                          title="ECDF",
-                          output.filename=NULL,
+                          title = "ECDF",
+                          output.filename = NULL,
                           palette = SeedMatchR.palette,
-                          factor.order=NULL,
+                          factor.order = NULL,
                           x.lims = c(-1,1),
                           stats.test = NULL,
-                          nboots = 10000,
+                          alternative = "greater",
                           null.name = 1,
                           target.name = 2,
-                          keep.boots = TRUE,
-                          plot.hist = TRUE,
                           height = 5,
-                          width=10,
-                          dpi=320){
+                          width = 5,
+                          dpi = 320){
   # This is to prevent a no visible binding for global variable error
   log2FoldChange <- Group <- NULL
 
@@ -136,8 +128,7 @@ deseq_fc_ecdf <- function(res,
                                 gene.lists[null.name],
                                 gene.lists[target.name],
                                 stats.test = stats.test,
-                                nboots = nboots,
-                                keep.boots = keep.boots)
+                                alternative = alternative)
 
   # Create ggplot
   ecdf.plot = ggplot2::ggplot(as.data.frame(res),
@@ -150,33 +141,18 @@ deseq_fc_ecdf <- function(res,
     ggplot2::labs(title = title,
                   x = "Log2(Fold Change)",
                   y="Cumulative Distribution",
-                  subtitle = ifelse(!is.null(stats.test),
-                                    paste0("Dstat: ",
-                                           round(test.results["Test Stat"],
-                                                 3),
-                                           "\n",
-                                           "Pvalue: ",
-                                           test.results["P-Value"]),
-                                    "No test stats")) +
+                  subtitle = paste0("Dstat: ",
+                                    test.results$statistic,
+                                    "\n",
+                                    "Pvalue: ",
+                                    test.results$p.value)) +
     SeedMatchR.theme +
     ggplot2::coord_cartesian(xlim = x.lims)
-
-  # If plot.hist is true, create a two column plot with the ecdf + hist
-  if (plot.hist == TRUE){
-    testit::assert("If plot.hist == TRUE,then keep.boots must also be TRUE",
-                   keep.boots == TRUE)
-    hist.plot = dstat_histogram(test.results)
-    #final.plot <- ecdf.plot + hist.plot
-    final.plot <- cowplot::plot_grid(ecdf.plot, hist.plot)
-
-  } else {
-    final.plot <- ecdf.plot
-  }
 
   # If there is an output name provided, save the figure
   if (!is.null(output.filename)){
     ggplot2::ggsave(output.filename,
-                    final.plot,
+                    ecdf.plot,
                     device="pdf",
                     height=height,
                     width=width,
@@ -184,5 +160,5 @@ deseq_fc_ecdf <- function(res,
                     dpi=dpi)
   }
 
-  return(list(plot = final.plot, stats = test.results))
+  return(list(plot = ecdf.plot, stats = test.results))
 }