+++ R/load_example_data.R
@@ -9,8 +9,6 @@
 load_example_data <- function(example.type){
   data.path = tempdir()
 
-  message(paste0("Example data directory being created at: ", data.path))
-
   sirna.path = paste0(data.path, "/Schlegel_2022.Rdata")
   mirna.path = paste0(data.path, "/mirdb.Rdata")